refactor: 폴더 구조 변경 및 라우터 경로 변경 #4
