+++ src/components/admin/AppSlider.vue
